+++ Demo/Sources/Demos/Demo Screens/FlowLayoutViewController.swift
@@ -2,155 +2,140 @@
 //  FlowLayoutViewController.swift
 //  Demo
 //
-//  Created by Kyle Van Essen on 1/13/20.
-//  Copyright © 2020 Kyle Van Essen. All rights reserved.
+//  Created by Kyle Van Essen on 11/25/21.
+//  Copyright © 2021 Kyle Van Essen. All rights reserved.
 //
 
-import UIKit
+@testable import ListableUI
 
-/**
- Implements  a very basic UICollectionViewFlowLayout, so we can use the debugger and view inspector
- to determine how it implements various things like supplementary views, etc.
- */
-final class FlowLayoutViewController : UIViewController
+import BlueprintUILists
+import BlueprintUICommonControls
+
+
+final class FlowLayoutViewController : ListViewController
 {
-    let layout : UICollectionViewFlowLayout
-    let collectionView : UICollectionView
+    var isHorizontal : Bool = false
     
-    override init(nibName nibNameOrNil: String?, bundle nibBundleOrNil: Bundle?)
-    {
-        self.layout = UICollectionViewFlowLayout()
-        self.collectionView = UICollectionView(frame: .zero, collectionViewLayout: self.layout)
+    override func viewDidLoad() {
+        super.viewDidLoad()
         
-        super.init(nibName: nibNameOrNil, bundle: nibBundleOrNil)
-        
-        self.title = "Flow Layout Demo"
+        self.navigationItem.rightBarButtonItem = UIBarButtonItem(
+            title: "Toggle Direction",
+            style: .plain,
+            target: self,
+            action: #selector(toggleDirection)
+        )
     }
     
-    required init?(coder: NSCoder) { fatalError() }
-    
-    override func loadView()
-    {
-        self.view = self.collectionView
+    @objc private func toggleDirection() {
+        self.isHorizontal.toggle()
         
-        self.collectionView.backgroundColor = .white
-        
-        self.collectionView.register(FlowCell.self, forCellWithReuseIdentifier: "Cell")
-        self.collectionView.register(FlowHeader.self, forSupplementaryViewOfKind:UICollectionView.elementKindSectionHeader, withReuseIdentifier: "Header")
-        self.collectionView.register(FlowFooter.self, forSupplementaryViewOfKind:UICollectionView.elementKindSectionFooter , withReuseIdentifier: "Footer")
-        
-        self.layout.headerReferenceSize = CGSize(width: 300.0, height: 50.0)
-        self.layout.footerReferenceSize = CGSize(width: 300.0, height: 50.0)
-        self.layout.itemSize = CGSize(width: 300.0, height: 100.0)
-        
-        self.layout.minimumLineSpacing = 20.0
-        self.layout.minimumInteritemSpacing = 20.0
-                
-        self.collectionView.delegate = self
-        self.collectionView.dataSource = self
+        self.reload(animated: true)
     }
     
-    fileprivate let items : [[FlowItem]] = [
-        [
-            // Empty section.
-        ],
-        [
-            // Section with one item.
+    override func configure(list: inout ListProperties) {
+        
+        list.appearance = .demoAppearance
+        
+        list.layout = .flow { flow in
+                        
+            flow.direction = isHorizontal ? .horizontal : .vertical
             
-            FlowItem(title: "Item 0, Section 1")
-        ],
-        [
-            // Section with two items.
+            flow.bounds = .init(
+                padding: UIEdgeInsets(top: 20, left: 20, bottom: 20, right: 20),
+                width: .atMost(900)
+            )
             
-            FlowItem(title: "Item 0, Section 2"),
-            FlowItem(title: "Item 1, Section 2")
-        ],
-        [
-            // Section with three items.
-            
-            FlowItem(title: "Item 0, Section 3"),
-            FlowItem(title: "Item 1, Section 3"),
-            FlowItem(title: "Item 2, Section 3")
+            flow.spacings = .init(
+                headerToFirstSectionSpacing: 20,
+                interSectionSpacing: .init(30),
+                sectionHeaderBottomSpacing: 10,
+                itemSpacing: 10,
+                rowSpacing: 10,
+                rowToSectionFooterSpacing: 10,
+                lastSectionToFooterSpacing: 20
+            )
+        }
+        
+        list.header = DemoHeader(title: "Flow Layout Demo")
+        list.footer = DemoHeader(title: "Flow Layout Footer")
+        
+        let ipsums : [String] = [
+            "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Vivamus non efficitur elit",
+            "Maecenas at ligula vitae nisl blandit pulvinar id at nibh. Mauris consequat vitae turpis id finibus. Nam eget massa ac augue commodo auctor",
+            "Morbi metus urna, ullamcorper in dapibus quis, ultricies id sem. Donec sodales efficitur odio, id pharetra ligula ornare vitae. Praesent lacinia mollis ipsum viverra molestie. Lorem ipsum dolor sit amet, consectetur adipiscing elit.",
+            "Mauris a nisi efficitur sem aliquet efficitur. Aenean nec mollis turpis. Integer erat risus, laoreet tincidunt aliquet quis, vestibulum non felis.",
+            "Donec urna magna, egestas et eros eu, pellentesque fringilla dui.",
+            "Nunc tempor interdum lectus, a pellentesque lorem consectetur non. Maecenas ullamcorper dapibus odio interdum ullamcorper.",
+            "Fusce aliquam tortor sit amet neque consequat faucibus. Nulla sem lorem, vehicula eget porta vitae, ultricies et sapien.",
+            "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Vivamus non efficitur elit",
         ]
-    ]
-}
-
-
-extension FlowLayoutViewController : UICollectionViewDataSource, UICollectionViewDelegate, UICollectionViewDelegateFlowLayout
-{
-    // MARK: UICollectionViewDataSource
-    
-    func numberOfSections(in collectionView: UICollectionView) -> Int
-    {
-        return self.items.count
-    }
-    
-    func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int
-    {
-        return self.items[section].count
-    }
-    
-    func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell
-    {
-        return collectionView.dequeueReusableCell(withReuseIdentifier: "Cell", for: indexPath)
-    }
-    
-    func collectionView(_ collectionView: UICollectionView, viewForSupplementaryElementOfKind kind: String, at indexPath: IndexPath) -> UICollectionReusableView
-    {
-        switch kind {
-        case UICollectionView.elementKindSectionHeader:
-            return collectionView.dequeueReusableSupplementaryView(ofKind: kind, withReuseIdentifier: "Header", for: indexPath)
-        case UICollectionView.elementKindSectionFooter:
-            return collectionView.dequeueReusableSupplementaryView(ofKind: kind, withReuseIdentifier: "Footer", for: indexPath)
-        default: fatalError()
+        
+        list.add {
+            for (alignment, description) in FlowAppearance.RowItemsAlignment.allTestCases {
+                Section("\(alignment)") { section in
+                    
+                    section.header = DemoHeader(
+                        title: description,
+                        detail: """
+                        The item alignment controls what happens when the items in the current \
+                        row are not the same height, allowing you to control the vertical alignment.
+                        """,
+                        useMonospacedTitleFont: true
+                    )
+                    
+                    section.layouts.flow.rowItemsAlignment = alignment
+                    section.layouts.flow.itemSizing = .columns(3)
+                    
+                    section.add {
+                        for ipsum in ipsums {
+                            FlowItem(text: ipsum)
+                        }
+                    }
+                }
+            }
+            
+            for (underflow, description) in FlowAppearance.RowUnderflowAlignment.allTestCases {
+                Section("\(underflow)") { section in
+                    
+                    section.header = DemoHeader(
+                        title: description,
+                        detail: """
+                        The underflow alignment controls what happens when the items in the current \
+                        row do not take up the full width of the layout; allowing you to align the results.
+                        """,
+                        useMonospacedTitleFont: true
+                    )
+                    
+                    section.layouts.flow.rowUnderflowAlignment = underflow
+                    section.layouts.flow.itemSizing = .fixed(200)
+                    
+                    section.add {
+                        for ipsum in ipsums {
+                            FlowItem(text: ipsum)
+                        }
+                    }
+                }
+            }
         }
     }
-    
-    // MARK: UICollectionViewDelegate
-    
-    // MARK: UICollectionViewDelegateFlowLayout
 }
 
 
-fileprivate struct FlowItem : Equatable
-{
-    var title : String
-}
-
-
-final fileprivate class FlowCell : UICollectionViewCell
-{
-    override init(frame: CGRect)
-    {
-        super.init(frame: frame)
-        
-        self.backgroundColor = .init(white: 0.9, alpha: 1.0)
+fileprivate struct FlowItem : BlueprintItemContent, Equatable {
+    
+    var text : String
+    
+    var identifierValue : String {
+        text
     }
     
-    required init?(coder: NSCoder) { fatalError() }
-}
-
-
-final fileprivate class FlowHeader : UICollectionReusableView
-{
-    override init(frame: CGRect)
-    {
-        super.init(frame: frame)
-        
-        self.backgroundColor = .init(white: 0.95, alpha: 1.0)
+    func element(with info: ApplyItemContentInfo) -> Element {
+        Label(text: text) {
+            $0.numberOfLines = 0
+            $0.color = .darkGray
+        }
+        .inset(uniform: 20.0)
+        .box(background: .white, corners: .rounded(radius: 10))
     }
-    
-    required init?(coder: NSCoder) { fatalError() }
-}
-
-final fileprivate class FlowFooter : UICollectionReusableView
-{
-    override init(frame: CGRect)
-    {
-        super.init(frame: frame)
-        
-        self.backgroundColor = .init(white: 0.8, alpha: 1.0)
-    }
-    
-    required init?(coder: NSCoder) { fatalError() }
 }